cleaned up cost functions, final testcases commit
