+++ py_valuenormalization/data/sample_test.txt
@@ -1,0 +1,5 @@
+University
+Univ
+Un
+UW madison
+UW
