+++ py_valuenormalization/tests/sample_test.txt
@@ -1,0 +1,5 @@
+University
+Univ
+Un
+UW madison
+UW
