+++ py_valuenormalization/tests/sample_test2.txt
@@ -1,0 +1,2 @@
+University	Wisc
+Johnson	street
